fix: run forge fmt

+++ src/PriceConverter.sol
@@ -5,16 +5,12 @@ import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/shared/interf
 
 library PriceConverter {
     function getPrice() internal view returns (uint256) {
-        AggregatorV3Interface priceFeed = AggregatorV3Interface(
-            0x694AA1769357215DE4FAC081bf1f309aDC325306
-        );
-        (, int answer, , , ) = priceFeed.latestRoundData();
-        return uint(answer) * 1e10;
+        AggregatorV3Interface priceFeed = AggregatorV3Interface(0x694AA1769357215DE4FAC081bf1f309aDC325306);
+        (, int256 answer,,,) = priceFeed.latestRoundData();
+        return uint256(answer) * 1e10;
     }
 
-    function getConversionRate(
-        uint256 ethAmount
-    ) internal view returns (uint256) {
+    function getConversionRate(uint256 ethAmount) internal view returns (uint256) {
         uint256 ethPrice = getPrice();
         uint256 ethAmountInUsd = (ethPrice * ethAmount) / 1e18;
         return ethAmountInUsd;
